SemanticPlag-main/requirements.txt

# Core ML / NLP
numpy>=1.24.0,<3.0.0
scikit-learn>=1.3.0
sentence-transformers>=2.2.0
torch>=2.0.0

# NLP preprocessing (NLTK only — beginner-friendly; no spaCy model download required)
nltk>=3.8.0

# Google Agent Development Kit — deterministic workflow orchestration
google-adk>=1.0.0
# Optional: uncomment for bonus features
# faiss-cpu>=1.7.4
# streamlit>=1.28.0
# reportlab>=4.0.0